Changed several base type names
git-svn-id: https://svn.vis.uni-stuttgart.de/utilities/vislib/trunk@1279 3ec011a9-ccf5-4689-8f2f-5f4aa021d377

+++ d3d/src/AbstractD3D11WindowImpl.cpp
@@ -234,7 +234,7 @@ void vislib::graphics::d3d::AbstractD3D11WindowImpl::initialise(HWND hWnd) {
                 NULL,
                 flags,
                 !featureLevels.IsEmpty() ? featureLevels.PeekElements() : NULL,
-                featureLevels.Count(),
+                static_cast<UINT>(featureLevels.Count()),
                 D3D11_SDK_VERSION,
                 &this->device,
                 &featureLevel,
